DRY diagram_generation_helper.sh to centralize tests definitions
With this change, adding a test only requires adding a single line of code defining a unique test name and the corresponding params.

It's a bit convoluted because associative arrays in bash don't keep the order so I have to keep a separate ordered array with the names to loop in order later. But I think it's clear-enough.